dataset/SkillTrustBench/secureSkills/case_01533/requirements.txt
requests>=2.28.0
aiohttp>=3.8.0
PyYAML>=6.0
